Refactor MigrationService: Consolidate services into single facade implementation
